Moved the Xilinx and Questa specific tool and library locations to a central place for the DRL.
- The Make flow (which I like) for Questa and the bash scripted Vivado
flow (which would be really slow if I had to open Vivado for every IP
run) share some common settings, so I pulled them out of the makefile
and made a tools.conf file that should contain everything specific to
the tool location and version and so forth.

- This assumes that the user will set up their environment to run with
the indicated Questa tools. The Makefile still assumes you are using the
appropriate versions