Finally got the interface to the UART wired up correctly
- AXI4-Lite interface was missing the write strobe signals (modport saved me
there)
- Master side of the interface is connected to the TB directly, but had to add a
wrapper around the VHDL UART, which is fine - couple hours of head banging to
get the interface with modports with clocking blocks to work properly without
complaining about multiply driven nets.
- Signals that are in a clocking block HAVE to go through the `cb_slave` or
whatever to be accessed. The clocking block constrains them that way (it's
almost like another module in the hierarchy)