Wow that was so much indirection I can barely keep my head straight
- Now have a testbench package to import that contains the uart_env, uart_test,
and probably the monitor and sequencer later (whatever those are)
- The whole thing works now with having removed modports. Still a lingering
thing I don't understand about clocking and output vs inout ports, but I'll sort
that out later today.
- I do not understand all the abstraction yet, but I'm reasonably sure that this
is correct.
- Basically implementing a UVM style testbench but without any of the existing
framework for reasons of complexity and licensing (it uses assertions and more
high-zoot simulator features that I do not have access to).