Giant fucking refactoring turned in revising the entire tree
- SystemVerilog doesn't support nested packages, so you need to take the UVM
route and just use a more switched on way of assembling source code
- I'm glad I'm the only that has to read this commit message
- The plan is to use a nested sequences of include files and relative paths so
that my future self only has to something like

```
`include "verif_macros.svh"
import verif_pkg::*;
```

And then in the Makefile that wants to use this, you just build with the
Makefile fragment that is in the root directory and pass `+incdir+$(VERIF_ROOT)`
to the command that compiles the verification framework package.

No muss, no fuss...Two lines.